% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{format_numbers}
\alias{format_numbers}
\title{Format Numbers for Plotting}
\usage{
format_numbers(x, digits = 2, scientific = FALSE, big_mark = ",")
}
\arguments{
\item{x}{Numeric vector to format}

\item{digits}{Number of digits to round to (default: 2)}

\item{scientific}{Whether to use scientific notation (default: FALSE)}

\item{big_mark}{Character to use for thousands separator (default: ",")}
}
\value{
Character vector of formatted numbers
}
\description{
A helper function to format numbers for use in plots, with options for
rounding, scientific notation, and adding suffixes.
}
\examples{
format_numbers(c(1234.567, 8901.234))
format_numbers(c(1234567, 8901234), scientific = TRUE)
}
