% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{\%notin\%}
\alias{\%notin\%}
\title{Not In Operator}
\usage{
x \%notin\% y
}
\arguments{
\item{x}{Vector of values to check}

\item{y}{Vector to check against}
}
\value{
Logical vector
}
\description{
A helper operator for checking if values are NOT in a vector.
This is the complement of the \%in\% operator.
}
\examples{
1:5 \%notin\% c(3, 4, 5)
}
