% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{percent}
\alias{percent}
\title{Percentage Calculator}
\usage{
percent(x, total, digits = 1)
}
\arguments{
\item{x}{Numerator value(s)}

\item{total}{Denominator value}

\item{digits}{Number of digits to round to (default: 1)}
}
\value{
Numeric vector of percentages
}
\description{
Calculate percentages with optional rounding.
}
\examples{
percent(25, 100)
percent(c(10, 20, 30), 100)
}
